refactor: Menu 패키징 작업
