+++ src/main/java/kitchenpos/service/menu/MenuGroupService.java
@@ -1,9 +1,9 @@
-package kitchenpos.service;
+package kitchenpos.service.menu;
 
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuGroupRepository;
-import kitchenpos.dto.MenuGroupRequest;
-import kitchenpos.dto.MenuGroupResponse;
+import kitchenpos.dto.menu.MenuGroupRequest;
+import kitchenpos.dto.menu.MenuGroupResponse;
 
 import java.util.List;
